Change for long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[]{5, 3, 6, 2});
+        SalesManager salesManager = new SalesManager(new long[]{5L, 3L, 6L, 2L});
 
         System.out.println(salesManager.max());
     }
